refactor(imports): replace design-system imports with utils for consistent utility usage across components
