+++ components/marketing/animated-word-swap.component.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { cn } from '@/lib/design-system';
+import { cn } from '@/lib/utils';
 
 /**
  * Props for the AnimatedWordSwap component
