+++ components/marketing/floating-card.component.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useRef, type ReactNode } from 'react';
-import { cn } from '@/lib/design-system';
+import { cn } from '@/lib/utils';
 
 /**
  * Props for the FloatingCard component
