+++ components/payments/pricing-toggle.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { cn } from '@/lib/design-system';
+import { cn } from '@/lib/utils';
 
 export type BillingInterval = 'month' | 'year';
